all: rename the tool to hey, add minor fixes
DO NOT MERGE

It was a mistake that this tool has been published with the same
name as tarekziade's boom. To avoid the conflict, we should rename the
tool at the cost of breaking the existing users.

Fixes #130.
